Use tabs instead of spaces in pseudo-tag lines.
This is to meet the specification in http://ctags.sourceforge.net/FORMAT,
according to which columns have to be tab-separated.

We already tab-separated normal tag lines, but the pseudo-tag lines
in the header were accidentally separated by spaces.

This fixes problems in some ctags file format parsers
(in my case, Sublime Text's CTags plugin).